Remove `sbom` extension (#31)
The Phylum UI now has the ability to export SBOMs from a submitted job.
It can do this in both SPDX and CycloneDX formats. The `sbom` community
extension can only provide the benefit of not needing to go through the
UI to get the SBOM. However, the `sbom` extension has not been updated
through the various CLI extension API breaking changes that have
happened over the past year. Therefore, it appears this extension is not
used and can be removed.